requirements.txt
alpaca_trade_api==3.2.0
kafka-python==2.0.2
nltk==3.9.1
alpaca-py==0.31.0
timeframe==3.0.0